Pdf file with different paths
This pdf contains examples of paths calculated with different algorithms to and from Sapo and Merinel caves. The set of sites is random and has been created to protect the exact location of the sites due to legal reasons. Maps in odd pages refer to Sapo cave paths and the ones in even pages refer to Merinel cave. Those path can be reproduce using the [script](paths2sapomerinel.R),